// src/app/types/documents/documents.enums.ts
export enum DocumentChangeTypeEnum {
  ADDED = 'added',
  REMOVED = 'removed',
  MODIFIED = 'modified',
}

export enum DocumentStatusEnum {
  DRAFT = 'draft',
  REVIEW = 'review',
  FINAL = 'final',
}

export enum DocumentCategoryEnum {
  LEGAL = 'legal',
  CONTRACT = 'contract',
  ADMINISTRATIVE = 'administrative',
}